third_party/blink/web_tests/wpt_internal/css/css-anchor-position/position-fallback-interpolation.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Anchor Positioning Test: Interpolation of position-fallback-bounds</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/9868">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>
<body>
<script>
  test_no_interpolation({
    property: 'position-fallback-bounds',
    from: 'normal',
    to: '--foo',
  });
  test_no_interpolation({
    property: 'position-fallback-bounds',
    from: '--foo',
    to: '--bar',
  });
</script>
